wsserver/requirements.txt
virtualenv
websocket-server
websocket-client
